Create HW2_PT2
Given an alphabet, print all the alphabets up to and including the given alphabet.
 
Write a function: 
        void printAlphabet(char ch) 
 
that accepts a character ch, and prints alphabets up to the given character. 
 
Input
    d 
 
Output
    a b c d 
 
    Where, 
There must be a single space between consecutive alphabets.
There should be no spaces after the last alphabet.
 
ch is a character within the range [{'a'...'z'}, {'A'...'Z'} ]. 
Alphabet(s) in input & output are case-sensitive.
NO output for any other input.